@@ -1,14 +1,14 @@
 name: whatsapp_stickers_example_dynamic
 description: Demonstrates how to use the whatsapp_stickers plugin with dynamic content.
-publish_to: 'none'
+publish_to: "none"
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  archive: ^2.0.11
+  archive: ^3.1.2
 
 dev_dependencies:
   flutter_test:
