@@ -3,23 +3,18 @@ import 'package:flutter/material.dart';
 import 'package:flutter_whatsapp_stickers/flutter_whatsapp_stickers.dart';
 
 Future<void> processResponse(
-    {StickerPackResult action,
-    bool result,
-    String error,
-    BuildContext context,
-    Function successCallback}) async {
-  print("_listener");
-  print(action);
-  print(result);
-  print(error);
-
-  SnackBar snackBar;
+    {StickerPackResult action = StickerPackResult.UNKNOWN,
+    bool? result,
+    String error = "Error",
+    BuildContext? context,
+    Function? successCallback}) async {
+  SnackBar? snackBar;
 
   switch (action) {
     case StickerPackResult.SUCCESS:
     case StickerPackResult.ADD_SUCCESSFUL:
     case StickerPackResult.ALREADY_ADDED:
-      successCallback();
+      successCallback!();
       break;
     case StickerPackResult.CANCELLED:
       snackBar = SnackBar(content: Text('Cancelled Sticker Pack Install'));
@@ -30,10 +25,12 @@ Future<void> processResponse(
     case StickerPackResult.UNKNOWN:
       snackBar = SnackBar(content: Text('Unkown Error - check the logs'));
       break;
+    default:
+      snackBar = SnackBar(content: Text('Unkown Error - check the logs'));
+      break;
   }
 
-  /// Display a snack bar
   if (snackBar != null && context != null) {
-    Scaffold.of(context).showSnackBar(snackBar);
+    ScaffoldMessenger.of(context).showSnackBar(snackBar);
   }
 }
